Get Latest Code to Fork

+++ src/core/thread/child_table.cpp
@@ -105,7 +105,7 @@ ChildTable::ChildTable(Instance &aInstance)
     memset(mChildren, 0, sizeof(mChildren));
 }
 
-Child *ChildTable::GetChildAtIndex(uint8_t aChildIndex)
+Child *ChildTable::GetChildAtIndex(uint16_t aChildIndex)
 {
     Child *child = NULL;
 
@@ -209,9 +209,9 @@ exit:
     return rval;
 }
 
-uint8_t ChildTable::GetNumChildren(StateFilter aFilter) const
+uint16_t ChildTable::GetNumChildren(StateFilter aFilter) const
 {
-    uint8_t      numChildren = 0;
+    uint16_t     numChildren = 0;
     const Child *child       = mChildren;
 
     for (uint16_t num = mMaxChildrenAllowed; num != 0; num--, child++)
@@ -225,7 +225,7 @@ uint8_t ChildTable::GetNumChildren(StateFilter aFilter) const
     return numChildren;
 }
 
-otError ChildTable::SetMaxChildrenAllowed(uint8_t aMaxChildren)
+otError ChildTable::SetMaxChildrenAllowed(uint16_t aMaxChildren)
 {
     otError error = OT_ERROR_NONE;
