+++ src/core/thread/mesh_forwarder.cpp
@@ -797,7 +797,7 @@ start:
             fragmentHeader = reinterpret_cast<Lowpan::FragmentHeader *>(payload);
             fragmentHeader->Init();
             fragmentHeader->SetDatagramSize(aMessage.GetLength());
-            fragmentHeader->SetDatagramTag(aMessage.GetDatagramTag());
+            fragmentHeader->SetDatagramTag(static_cast<uint16_t>(aMessage.GetDatagramTag()));
             fragmentHeader->SetDatagramOffset(0);
 
             payload += fragmentHeader->GetHeaderLength();
@@ -824,7 +824,7 @@ start:
         fragmentHeader = reinterpret_cast<Lowpan::FragmentHeader *>(payload);
         fragmentHeader->Init();
         fragmentHeader->SetDatagramSize(aMessage.GetLength());
-        fragmentHeader->SetDatagramTag(aMessage.GetDatagramTag());
+        fragmentHeader->SetDatagramTag(static_cast<uint16_t>(aMessage.GetDatagramTag()));
         fragmentHeader->SetDatagramOffset(aMessage.GetOffset());
 
         payload += fragmentHeader->GetHeaderLength();
